Rename 'misc' methods and subdirectory to 'results'.
Rename the 'misc' subdirectory and methods to 'results' to better
reflect the meaning of these files.  The 'misc' files are additional
results to the retrace_backtrace for the task, which are basically
analysis output files generated from output of a debugger command
on the corefile or vmcore.  This directory will be given to 'hooks'
scripts as a place to write output files.

Note that this change will handle existing tasks with the 'misc'
directory in a transparent manner, but new tasks will write to
and read from a new 'results' directory.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>