Handle 'flattened' vmcore format by converting with makedumpfile
The 'flattened' format of vmcore files should be converted to a
'normal' dumpfile format with makedumpfile -R before the crash-utility
is run.  With a non-converted file, getting to the crash-utility
command prompt may take many minutes.  In our production system it is
not uncommon to see a few vmcores take up to an hour get to the
crash-utility command prompt.

The 'flattened' format is not widely known by engineers that analyze
vmcores, nor is it known that this format can be easily converted with
makedumpfile.  The conversion is something that can easily be handled
by retrace-server as part of vmcore setup.

Add a KernelVMcore class that can handle both identifying and
converting the 'flattened' format.  This class should be useful in the
future, and functions such as get_vmcore_dump_level, strip_vmcore,
and get_kernel_release could be moved into this class.

When the 'flattened' file is detected, log a message regarding
makedumpfile execution similar to the message when makedumpfile is run
to strip extra pages.  In addition, slightly refactor the existing
makedumpfile message involving stripped pages to be consistent with
the makedumpfile execution for converting 'flattened' files.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1753635

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>