Add KernelVMcore.has_extra_pages method
Move the logic to determine whether a vmcore has pages that can be
stripped into a separate method.  The method returns true if the
vmcore has extra pages that can be stripped with makedumpfile, and
the return value is the inverted logic of the previous
'skip_makedumpfile' variable.  Should be no functional change.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>