Move get_kernel_release to KernelVMcore
Move get_kernel_release inside KernelVMcore as well as clean up
the open and read of the file and change OSerror to IOerror
which is the exception much more likely to occur (for example
if the vmcore does not exist or cannot be opened for some
reason).  In the process of moving this code, we must modify
the callers of prepare_debuginfo to make sure a KernelVMcore
object exists prior to calling with this object rather than the
vmcore_path.  Also add a get_path method to KernelVMcore to be
used by prepare_debuginfo which is not yet moved into
KernelVMcore class and yet needs the vmcore_path at a few
places.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>