Remove call to prepare_debuginfo from retrace-server-interact
The retrace-server-interact command should not be calling
prepare_debuginfo.  The only conditions when this might happen would
be if a task is still in progress or hung.  In both instances we
should not proceed but print an error message for the user on the
condition of the task and possible corrective actions.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>