Move makedumpfile logic from download_remote into start_vmcore
For a long time now the logic to scan for a kernel version, setup
the debuginfo symbols, and call makedumpfile have been in a difficult
to follow logic flow.  For example, in the case of vmcores that
needed stripped pages, we would call prepare_debuginfo from that
function, but then later call it again when we were setting up
the backtrace.  Clean all this up by moving the makedumpfile
logic and call to prepare_debuginfo inside start_vmcore.  Also add
_vmlinux internal variable to KernelVMcore which is set when we
are sure prepare_debuginfo has been successful.  In the process,
we can clean up strip_extra_pages to not need any parameters other
than the vmcore itself.

This fixes numerous problems where prepare_debuginfo might have been
called multiple times depending on the type of vmcore that was
submitted.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>