retrace: Keep suffixes of snapshots
One of the requirements for processing vmware snapshots with crash is
that the file extensions should NOT be changed.

But in retrace we take provided core file and rename it according to the
task type (coredump x vmcore). This would cause crash tool to fail to
process the vmware snapshot files (.vmsn, .vmss) and find a memory snapshot (.vmem).

With this commit we will keep file extensions for defined list of
suffixes (.vmsn, .vmss, .vmem).

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>