Fix manager page ftp submission due to missing check for values
Commit 53fd9b5 reworked the parsing of manager page and missed
a subtle change for null values.  By default the 'start' method
passes some of the parameters such as 'kernelver' with null values:
start?kernelver=&caseno=&bugzillano=&notify=&debug=on&md5sum=on

This results in the following error message if 'kernelver' is not
filled in:
Please use VRA format for kernel version (e.g. 2.6.32-287.el6.x86_64)

Fix this by checking for null values in kernelver, caseno, and
notify.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>