reposync: Stop accessing cachedir attribute of Repo objects
The cachedir property of Repo in DNF was made “private” in
62fcedbf32b9f0419dc4fcb9620860d51f4ffb29, largely due to $REASONS.
Following some more development since then, there is simply no way to
access the cache directory for individual repositories anymore (we
actually have correct code already, but this particular bit was missed).
The solution is to grab the base for cache and concatenate it with the
ID of the repository.