Remove get_mock and set_mock and add "native" RetraceEnvironment
There is some legacy logic to set a different version of crash based
on whether a specific config variable is set.  This was used primary
on x86_64 to run a 32-bit version of crash when needed, rather than
using mock.

This logic is somewhat confusing and not used anymore so remove this.
Instead introduce a "native" option for RetraceEnvironment.  For
now this is only implemented for kernel vmcores not userspace
coredumps so throw an exception with a message inside run_gdb().

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>