retrace-server-task: Add option to restart an existing task
Historically, restarting a failed task was done with
'retrace-server-worker --restart'.  However, this command was
not designed to be executed by a normal user, since the task
executes as a normal user in the restart case, which is
different from normal new task creation.

Instead, use retrace-server-task to 'restart' an existing task
which mirrors how 'create' instantiates a new task.  By using
retrace-server-task which then calls through the HTTP interface,
the task operations execute as 'retrace' user.  This is both
a cleaner user interface, and avoids numerous problems with
user credentials such as storing task files with user
ownership as well as issues with extracting kernel-debuginfos.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>