manager: Prevent running tasks from being restarted
Its possible that a task is still running when a user tries to restart
it.  Prevent this situation and tell the user since two instances of
retrace-server-worker operating on the same task will interere with
one another.
$ retrace-server-task restart 841048690
Restarted task '841048690'
$ retrace-server-task restart 841048690
Status code 403 was returned with message: 'Task is still running, cannot restart'

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>