retrace-server-task: Handle remote authentication failure
If retracer-server-task cannot connect to the (default) localhost
server, it falls back to retrace.fedoraproject.org, but does not check
if it's able to connect there.

The problem is that it falls back to retrace.fp.org if the connection
fails for any reason. Specifically, it falls back even if authentication
fails on localhost.

Moreover, if retrace.fp.org is inacessible as well, we get an exception
later down the road when trying to send an HTTP request.

In this change, we check if the server responded HTTP 401 Unauthorized,
which should be the case if a valaid Kerberos ticket was expected by the
server but not sent by the client.

Resolves #367