retrace-server-task: Add email, caseno, bugzillano options for restart
Commit 6bf02a8 implemented a 'restart' option on the retrace-server-task
command, but did not implement all options which would be appropriate
for restart.  Add options options email, caseno, bugzillano.  Note that
currently restart of a task does not download or recalculate the md5sum,
so md5sum option is not added.

Since these options are nearly the same as when we 'start' a new task,
refactor the manager code on the backend to avoid parsing duplication.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>