manager: Add 'Restart task' button to restart a failed task
Add a 'Restart task' link on the manager task page for any failed
task.  Clicking the link will open a new page which allows entry
of parameters similar to the page when a new ftp task is created.
On the restart page, a user can input Kernel version, Case no.,
Bugzilla no., E-mail notification, as well as checkboxes for
md5sum and verbosity.

Create a few helper functions which return the form data for
the new 'restart' page, which is very similar to the 'start'
method page, with only a few exceptions:
- 'restart': display existing values for caseno, md5sum, etc
- 'start': allow selection of md5 calculation

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>