manager: reformat manager page to improve readability
The manager page is heavily used, yet many of the fields in the most
important locations on the page are unused.  Move less used fields
such as the out of the premium locations and replace with the more
important fields.

Remove the ftp list - it's slow and only around 10% of vmcores.
Move finished tasks to the left portion of the page.
Move available and running tasks to the right portion of the page.
Add border around "Create vmcore task" and "Create userspace task"
sections.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>