ftp/manager: Add link on manager page and create backlink on ftp page
We removed the list of files from the manager page, but some users
may still want to create a task from an FTP file.

On the top of the manager page, add back a link to the FTP page
so the users that still use this can get to it to create a task.
In the manager wsgi file, remove now unused code which implemented
the display of ftp files in place of available tasks.

On the ftp page, add a small bit of CSS code before the table to
better match what was on the manager page.  Also at the top of
the FTP page, add a link back to the task manager.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>