manager: Rework display of finished, running, and available tasks
There's no need to waste screen space for the running or available
tasks if there are none of these.  Rework the display of the tables
to be more dynamic, and only display header if there are rows in
the table.

In addition, make the Running and Available tasks display above
the Finished tasks, and take up 100% of the screen width.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>