util: Repair parsing regexes
The changes in 84d67af made the regexes accept more inputs as valid than
originally intended due to the use of PCRE shorthands (\w, \d, etc.).
This commit partially reverts to the original and adds the ASCII flag to
`re.compile()` to ensure only characters in the ASCII range are
accepted for the shorthands.