spec: Migrate old home directory in post scriptlet
The home directory of the retrace user was changed in 921759f but no
provisions were made to migrate the home directory on an old
installation to the new one. This change remediates this shortcoming.

In addition, add systemctl as a dependency since we need it to test if
httpd is running.