manager: Add existence and permissions check of custom_url to avoid failed tasks
Tasks that are created with custom_url of a local path that resolves to
either a non-existent file, or a file that exists but the retrace-server
processes cannot read will leave a failed task and delayed feedback to
the user.

Add an early check to open the file for reading inside the __custom__ URL
parsing inside the manager.wsgi file.  This allows the httpd process to
respond to the user with an appropriate error message, giving early
feedback of the problem and avoiding tasks that would eventually fail.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>