bugzilla-query: Use 'limit' and search only recently modified bugs
Set the 'limit' option to bzapi.build_query() function to ensure we
are not artificially limited to a small default number of bugs.
In production we were seeing an artificial limit of 20 bugs get
queried without setting this option.  If we ever meet our defined
limit of 500 bugs in the query, this means we probably did not
query all bugs.  If we hit the limit, then loop around until we
query all bugs by using the 'offset' option.

In addition, cut down on the number of bugs in the query by only
querying those with recent modification times.  Define a new
config variable, BugzillaQueryLastChangeDelta, which is the number
of hours ago we query for changed bugs.  By default set this to
168 hours (7 days), and with the understanding that the cronjob
that runs bugzilla-query should be run at least once per week.  We
can look at only the recently modified bugs because we are looking
for additions of retrace-server task strings in the bug comment
text.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>