Add logic to determine if we save a stripped vmcore
Currently we do not check to see if running makedumpfile and creating
a stripped vmcore is actually worthwhile. In some situations we may
see little to no benefit to stripping down a vmcore and replacing the
original with the newly stripped vmcore. For those situation, this
patch adds the config variable VmcoreDumpSavePercent which is meant
to indicate the percentage of space a stripped vmcore will need to
save for us to choose to save it in place of the original. The
proposed default is 10%. This patch also adjusts the logic so that
the relevant message about makedumpfile will be printed (i.e. if
makedumpfile is not run or we don't use the stripped vmcore from
makedumpfile we do not print how much space we saved).

Resolves #445
Signed-off-by: Audra Mitchell <aubaker@redhat.com>