retrace-server-interact: Fix traceback for non-existent gid
Its possible for a user to be in a group that does not have a name.
In this case retrace-server-interact will Traceback something like
the following
 Traceback (most recent call last):
   File "/usr/bin/retrace-server-interact", line 31, in <module>
     groups = [grp.getgrgid(g).gr_name for g in os.getgroups()]
   File "/usr/bin/retrace-server-interact", line 31, in <listcomp>
     groups = [grp.getgrgid(g).gr_name for g in os.getgroups()]
 KeyError: 'getgrgid(): gid not found: 12345'

Fix this by ensuring we don't assume the group database will
have an entry for each gid, in which case getgrgid() will raise
KeyError.  If this happens, catch the exception and just skip
over the gid.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>