config: Add support for RAR archives if the unar tool is installed
If 'unar' tool is installed we can extract 'rar' archives, which
is a rare, but not non-existent, archive type.

Make this an optional install for meson to prevent build failures.
If not installed, but a RAR archive is submitted, the task will
fail with something like the following in retrace_log:
 [D] File type: rar archive data, v4, os: win32
 [D] rar detected
 [W] [Errno 2] No such file or directory: 'unar': 'unar'

Resolves: https://github.com/abrt/retrace-server/issues/465
Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1457434
Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>