retrace_worker: Move --minimal option to end of crash commandline
The crash-utility does not care what position the '--minimal' option
is when invoking the program.  Move it to the end of the line so
we only have to define this one time, simplyfing the code and will
make a future patch easier when we try to handle other crash options.
No functional change.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>