retrace-server-task: Add exception handling for SSLError
It's not uncommon in production to have a certificate expire,
and when that happens we use the FAILBACK_SERVER, but this
is unlikely to be what the user wants.

This patch catches the error and gives the user possible next
steps to avoid the error, similar to earlier commit cf85a18c1b29.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>